Further simplify the full.jar generation (the hard portability way)
Don't run the jar at build time to ensure that the java searches the
libs where the cmake puts them. It was nice but it's hard to get that
script running on windows.

Instead, manually determine in both systems (java and cmake) where the
libs should be. Of course both definitions should match.

 - Allows to remove the java_bundle.sh pimple (that breaks on windows)
 - Users compiling simgrid will use the installed libs instead
 - None of the solutions will work for users using a faulty jar
 - Our testing infrastructure will catch the glitches:
   Each system displays the libs location in the build logs.