[tesh] exec the killer process after fork to make it work on windows
fork() is only emulated on windows, using a thread of the current
process. So if we don't exec away the killer process, the tesh process
will not end before the timeout, leading to loooooong testing phases.

This script is still in a rather sorry state, more cleanups would be
welcome.