Fix duplicate PTP messages when tracing/smpi/internals is enabled
When the tracing/smpi/internals option is not enabled, some PTP messages
are showed for send and receive operations, both on tracing and
replaying. When the option is enabled, additional messages are
generated, so duplicate messages for send and receive operations are
duplicated.

Initially I thought that the idea was to only show PTP messages when the
tracing/smpi/internals option is enabled. After talking to Arnaud it
seems that is not the case. So I changed the code so both funcionalities
are present: PTP messages for send and receive operations when the
option is disabled and all PTP messages when the option is enabled, as
expected.