Fix Java build
java CPU surf binding used to register itself to the list of models that
must be called through shareResources directly. But this was useless
because CPU models are invoked from the host models (leading to
dupplicated lines in the tesh file).

Now that it is not possible anymore to register to that list of models
(which was removed), we must fix the java CPU surf binding.