Make HostL07 behave more like the regular Host
The methods execute() and sleep() were implemented directly in the
host, and the virtual functions in the cpu were implemented with a
DIE_IMPOSSIBLE.

Instead, HostL07 is dispatching the execute() and sleep() calls that
it gets from the higher layers to its CpuL07, just as any Host.

The goal is to kill HostL07 and just use a regular Host dispatching
*every* call to its CPU and RoutingEdge. Then, Host will be made part
of the public interface (simgrid::Host will offer the interface of
S4U::Host with the content of surf::Host).