not all resources need a property set, only Host, Link and Storage
So dupplicate the property-related code in these classes, and remove
it from Resource. We may want to refactorize this code in a specific
class, but I'm not confident with multiple inheritance yet.

The reason of fixing that working feature is the bloat fight. A CPU or
RoutingEdge don't need no properties. The Resource class contains too
much things and should be split of.