unspecialize the trace setup for CPUs
Every Cpu class exposes the same interface to set a state_trace or a
speed_trace, and every code that sets it use the interface.

As a result, every code setting the traces of a CPU is very similar
and should be factorized. For now, it's spread in the whole class tree.

Actually, the right place to do it is in the xml parser itself. The
symbol trace_connect_list_host_speed and friends should have remained
private to the xml parser.

I'll do so tomorrow, and also deal with the setup of link traces.