sanitize the surf API
updateState() and friends don't need to have the current date passed
as an parameter: they can use surf_get_time() on the rare cases where
they need it.

This makes the surf_solve loop rather ugly as the code is not ready to
jump to the right time to handle events yet. Will soon be fixed.