[mc] Don't leave model-checked processes around
In some cases, we terminate the model-checker brutally inside of its
event loop instead of letting the event loop and the model checker
terminate cleanly. In this case, the model-checked process could still
be left alive.

Because the model-checker was still around, tesh was still waiting on
the pipe and ended-up timing out.

We now SIGKILL the model-checked process when exiting the
model-checker brutally.