Revert "[SMPI] Fix wrong return value in MPI_Testany."
This reverts commit efcf0fb465cf0465c462fc28a830ccd4d6b6b13c.

The reason I reverted again is that the standard does actually state

"The array may contain null or inactive handles. If the array contains
no active handles then the call returns immediately with flag = true,
index = MPI_UNDEFINED , and an empty status."

This is a corner case and was correctly handled here.