[TOOLS] Added .clang-format config
This is a configuration for clang-format-3.8 or later.
It will help us to have a unified coding style and will
tell us if some code is not well formatted.

This file is only the configuration and does not activate
any hooks. You need to do that yourself, if you want.