move the VM-related data out of MSG's private data for hosts
This is the beginning of turning the VMs into a separate plugin. I'm
sure that reconcentrating all VM code into a few files will greatly
simplify it.

This change is really messy because I did not want to be too intrusive
in msg_vm.c and had to expose much much of the VM internals for that.
It will be improved in the future (it cannot get any worse).

This reveals a little bug where the dirtyPages things of a VM that
runs on a host which turns down are not properly cleaned up. There is
nothing I can do right now as I don't have a onHostOff event or
something, so I simply spit a warning in this case. For now.