call updateRemaining when the action is suspended (fix #123)
if not, the LMM is giving the right results, but surf does not notice.

Thus the fix does not touch LMM, only surf_interface.cpp. Note that
suspend has the specificity that we cannot estimate when it will end.
Thus it does not seem appropriate to use the nextOccuringEventLazy
method which is essentially trying to find when will be the next step.
Instead I propose to call updateReaminingLazy from the suspend
function.