NS3 extensions are for netcards, not for hosts
We need to have it for every netcard, be it an host, an AS or an
host-less router.

It's cumbersome to put it onto hosts when possible and onto netcard
for routers and ASes, and then search first in the host and if not
found in the netcards.

Instead, only put it on netcards, and we're set.