MSG_process_ref/unref: allow to fiddle with the process refcounting
This is useful to avoid that a process gets deleted before we're done
with it. This is what happened in the process-join example, where we
were joining the process after its end. So we have to increase its
refcount while it's still alive to make sure that it does not get
destroyed too early.

This would be automated in C++ with RAII but we are not there yet.