partially address a bug in Java initialization
SimGrid flag on command line were consumed at C level but stayed in the
original Java String[] args. This could mess users' args[i] if SG flags
were not put at the end of the command line.

This patch removes the SG flags from the Java arguments. However, the
number of arguments REMAINS UNCHANGED.

It is then UNSAFE to test if args.length is greater than the number of
YOUR OWN ARGUMENTS. It might be if you have --log or --cfg flags in
the command line. BTW, it already was ...